Rearranged folder organization for better clarity
