+++ layout-next/src/app/(internal)/layout.tsx
@@ -1,0 +1,7 @@
+import PageTemplate from "@/components/template/Page";
+
+export default function Layout({children}: any) {
+    return (
+        <PageTemplate>{children}</PageTemplate>
+    )
+}
