+++ layout-next/src/app/(internal)/page/page.tsx
@@ -1,0 +1,6 @@
+
+export default function Page() {
+    return (
+        <div> Exemplo básico de uma página!!!</div>
+    )
+}
