+++ layout-next/src/app/(internal)/second/page.tsx
@@ -1,9 +1,7 @@
 import Box from "@/components/Box";
-import PageTemplate from "@/components/template/Page";
 
 export default function Flexbox () {
     return (
-        <PageTemplate>
             <div className="h-96">
                 <h1>Flexbox</h1>
                 <div className="flex h-full w-[500px] justify-around items-end gap-2">
@@ -12,6 +10,5 @@ export default function Flexbox () {
                     <Box text="3"/>
                 </div>
             </div>
-        </PageTemplate>
     )
 }
