--- layout-next/src/app/page/page.tsx
@@ -1,9 +1,0 @@
-import PageTemplate from "@/components/template/Page";
-
-export default function Page() {
-    return (
-        <PageTemplate>
-            <div> Exemplo básico de uma página!!!</div>
-        </PageTemplate>
-    )
-}
